+++ src/CalcManager/Ratpack/itrans.cpp
@@ -92,7 +92,7 @@ void asinanglerat( _Inout_ PRAT *pa, ANGLE_TYPE angletype, uint32_t radix, int32
 void asinrat( PRAT *px, uint32_t radix, int32_t precision)
 
 {
-    long sgn;
+    int32_t sgn;
     PRAT pret= nullptr;
     PRAT phack= nullptr;
 
@@ -204,7 +204,7 @@ void _acosrat( PRAT *px, int32_t precision)
 void acosrat( PRAT *px, uint32_t radix, int32_t precision)
 
 {
-    long sgn;
+    int32_t sgn;
 
     sgn = (*px)->pp->sign*(*px)->pq->sign;
 
@@ -291,7 +291,7 @@ void _atanrat( PRAT *px, int32_t precision)
 void atanrat( PRAT *px, uint32_t radix, int32_t precision)
 
 {
-    long sgn;
+    int32_t sgn;
     PRAT tmpx= nullptr;
 
     sgn = (*px)->pp->sign * (*px)->pq->sign;
